+++ packages/tendermint-rpc/src/tendermint37/adaptor/index.ts
@@ -5,7 +5,7 @@ import { Adaptor } from "./types";
 
 export { Decoder, Encoder, Params, Responses } from "./types";
 
-export const adaptor35: Adaptor = {
+export const adaptor37: Adaptor = {
   params: Params,
   responses: Responses,
   hashTx: hashTx,