+++ packages/tendermint-rpc/src/tendermint37/tendermint37client.spec.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/naming-convention */
-import { toAscii } from "@cosmjs/encoding";
+import { toAscii, toHex } from "@cosmjs/encoding";
 import { firstEvent, toListPromise } from "@cosmjs/stream";
-import { sleep } from "@cosmjs/utils";
+import { assert, sleep } from "@cosmjs/utils";
 import { ReadonlyDate } from "readonly-date";
 import { Stream } from "xstream";
 
@@ -16,16 +16,16 @@ import {
   tendermintInstances,
   tendermintSearchIndexUpdated,
 } from "../testutil.spec";
-import { adaptor35 } from "./adaptor";
+import { adaptor37 } from "./adaptor";
 import { buildQuery } from "./requests";
 import * as responses from "./responses";
-import { Tendermint35Client } from "./tendermint35client";
+import { Tendermint37Client } from "./tendermint37client";
 
 function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues): void {
   describe("create", () => {
     it("can auto-discover Tendermint version and communicate", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
       const info = await client.abciInfo();
       expect(info).toBeTruthy();
       client.disconnect();
@@ -33,7 +33,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
 
     it("can connect to Tendermint with known version", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
       expect(await client.abciInfo()).toBeTruthy();
       client.disconnect();
     });
@@ -41,7 +41,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
 
   it("can get genesis", async () => {
     pendingWithoutTendermint();
-    const client = await Tendermint35Client.create(rpcFactory());
+    const client = await Tendermint37Client.create(rpcFactory());
     const genesis = await client.genesis();
     expect(genesis).toBeTruthy();
     client.disconnect();
@@ -50,7 +50,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
   describe("broadcastTxCommit", () => {
     it("can broadcast a transaction", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
       const tx = buildKvTx(randomString(), randomString());
 
       const response = await client.broadcastTxCommit({ tx: tx });
@@ -70,7 +70,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
   describe("broadcastTxSync", () => {
     it("can broadcast a transaction", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
       const tx = buildKvTx(randomString(), randomString());
 
       const response = await client.broadcastTxSync({ tx: tx });
@@ -86,7 +86,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
   describe("broadcastTxAsync", () => {
     it("can broadcast a transaction", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
       const tx = buildKvTx(randomString(), randomString());
 
       const response = await client.broadcastTxAsync({ tx: tx });
@@ -98,9 +98,9 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
 
   it("gets the same tx hash from backend as calculated locally", async () => {
     pendingWithoutTendermint();
-    const client = await Tendermint35Client.create(rpcFactory());
+    const client = await Tendermint37Client.create(rpcFactory());
     const tx = buildKvTx(randomString(), randomString());
-    const calculatedTxHash = adaptor35.hashTx(tx);
+    const calculatedTxHash = adaptor37.hashTx(tx);
 
     const response = await client.broadcastTxCommit({ tx: tx });
     expect(response.hash).toEqual(calculatedTxHash);
@@ -111,7 +111,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
   describe("abciQuery", () => {
     it("can query the state", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
 
       const key = randomString();
       const value = randomString();
@@ -137,7 +137,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
 
   it("can get a commit", async () => {
     pendingWithoutTendermint();
-    const client = await Tendermint35Client.create(rpcFactory());
+    const client = await Tendermint37Client.create(rpcFactory());
     const response = await client.commit(4);
 
     expect(response).toBeTruthy();
@@ -152,7 +152,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
 
   it("can get validators", async () => {
     pendingWithoutTendermint();
-    const client = await Tendermint35Client.create(rpcFactory());
+    const client = await Tendermint37Client.create(rpcFactory());
     const response = await client.validators({});
 
     expect(response).toBeTruthy();
@@ -170,7 +170,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
 
   it("can get all validators", async () => {
     pendingWithoutTendermint();
-    const client = await Tendermint35Client.create(rpcFactory());
+    const client = await Tendermint37Client.create(rpcFactory());
     const response = await client.validatorsAll();
 
     expect(response).toBeTruthy();
@@ -188,7 +188,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
 
   it("can call a bunch of methods", async () => {
     pendingWithoutTendermint();
-    const client = await Tendermint35Client.create(rpcFactory());
+    const client = await Tendermint37Client.create(rpcFactory());
 
     expect(await client.block()).toBeTruthy();
     expect(await client.genesis()).toBeTruthy();
@@ -200,7 +200,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
   describe("status", () => {
     it("works", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
 
       const status = await client.status();
 
@@ -230,7 +230,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
   describe("numUnconfirmedTxs", () => {
     it("works", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
 
       const response = await client.numUnconfirmedTxs();
 
@@ -244,7 +244,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
   describe("blockResults", () => {
     it("works", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
 
       const height = 3;
       const results = await client.blockResults(height);
@@ -260,7 +260,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
   describe("blockSearch", () => {
     beforeAll(async () => {
       if (tendermintEnabled()) {
-        const client = await Tendermint35Client.create(rpcFactory());
+        const client = await Tendermint37Client.create(rpcFactory());
 
         // eslint-disable-next-line no-inner-declarations
         async function sendTx(): Promise<void> {
@@ -285,7 +285,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
 
     it("can paginate over blockSearch results", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
 
       const query = buildQuery({ raw: "block.height >= 1 AND block.height <= 3" });
 
@@ -304,7 +304,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
 
     it("can get all search results in one call", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
 
       const query = buildQuery({ raw: "block.height >= 1 AND block.height <= 3" });
 
@@ -323,7 +323,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
   describe("blockchain", () => {
     it("returns latest in descending order by default", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
 
       // Run in parallel to increase chance there is no block between the calls
       const [status, blockchain] = await Promise.all([client.status(), client.blockchain()]);
@@ -340,7 +340,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
 
     it("can limit by maxHeight", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
 
       const height = (await client.status()).syncInfo.latestBlockHeight;
       const blockchain = await client.blockchain(undefined, height - 1);
@@ -354,7 +354,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
 
     it("works with maxHeight in the future", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
 
       const height = (await client.status()).syncInfo.latestBlockHeight;
       const blockchain = await client.blockchain(undefined, height + 20);
@@ -369,7 +369,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
 
     it("can limit by minHeight and maxHeight", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
 
       const height = (await client.status()).syncInfo.latestBlockHeight;
       const blockchain = await client.blockchain(height - 2, height - 1);
@@ -383,7 +383,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
 
     it("contains all the info", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
 
       const height = (await client.status()).syncInfo.latestBlockHeight;
       const blockchain = await client.blockchain(height - 1, height - 1);
@@ -412,7 +412,7 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
   describe("tx", () => {
     it("can query a tx properly", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
 
       const find = randomString();
       const me = randomString();
@@ -436,25 +436,6 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
       expect(r.height).toEqual(height);
       expect(r.proof).toBeTruthy();
 
-      // txSearch - you must enable the indexer when running
-      // tendermint, else you get empty results
-      const query = buildQuery({ tags: [{ key: "app.key", value: find }] });
-
-      const s = await client.txSearch({ query: query, page: 1, per_page: 30 });
-      // should find the tx
-      expect(s.totalCount).toEqual(1);
-      // should return same info as querying directly,
-      // except without the proof
-      expect(s.txs[0]).toEqual({ ...r, proof: undefined });
-
-      // ensure txSearchAll works as well
-      const sall = await client.txSearchAll({ query: query });
-      // should find the tx
-      expect(sall.totalCount).toEqual(1);
-      // should return same info as querying directly,
-      // except without the proof
-      expect(sall.txs[0]).toEqual({ ...r, proof: undefined });
-
       // and let's query the block itself to see this transaction
       const block = await client.block(height);
       expect(block.block.txs.length).toEqual(1);
@@ -465,25 +446,28 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
   });
 
   describe("txSearch", () => {
-    const key = randomString();
+    const txKey = randomString(); // a key used for multiple transactions
+    let tx1: Uint8Array | undefined;
+    let broadcast1: responses.BroadcastTxCommitResponse | undefined;
 
     beforeAll(async () => {
       if (tendermintEnabled()) {
-        const client = await Tendermint35Client.create(rpcFactory());
+        const client = await Tendermint37Client.create(rpcFactory());
 
         // eslint-disable-next-line no-inner-declarations
-        async function sendTx(): Promise<void> {
+        async function sendTx(): Promise<[Uint8Array, responses.BroadcastTxCommitResponse]> {
           const me = randomString();
-          const tx = buildKvTx(key, me);
+          const tx = buildKvTx(txKey, me);
 
           const txRes = await client.broadcastTxCommit({ tx: tx });
           expect(responses.broadcastTxCommitSuccess(txRes)).toEqual(true);
           expect(txRes.height).toBeTruthy();
-          expect(txRes.hash.length).not.toEqual(0);
+          expect(txRes.hash.length).toEqual(32);
+          return [tx, txRes];
         }
 
         // send 3 txs
-        await sendTx();
+        [tx1, broadcast1] = await sendTx();
         await sendTx();
         await sendTx();
 
@@ -493,22 +477,86 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
       }
     });
 
-    it("returns transactions in descending order by default", async () => {
+    it("finds a single tx by hash", async () => {
+      pendingWithoutTendermint();
+      assert(tx1 && broadcast1);
+      const client = await Tendermint37Client.create(rpcFactory());
+
+      const result = await client.txSearch({ query: `tx.hash='${toHex(broadcast1.hash)}'` });
+      expect(result.totalCount).toEqual(1);
+      expect(result.txs[0]).toEqual({
+        hash: broadcast1.hash,
+        height: broadcast1.height,
+        index: 0,
+        tx: tx1,
+        result: broadcast1.deliverTx!,
+        proof: undefined,
+      });
+
+      client.disconnect();
+    });
+
+    it("finds a single tx by tags", async () => {
+      pendingWithoutTendermint();
+      const client = await Tendermint37Client.create(rpcFactory());
+
+      const txKey2 = randomString();
+      const txValue2 = randomString();
+      const tx = buildKvTx(txKey2, txValue2);
+
+      const txRes = await client.broadcastTxCommit({ tx: tx });
+      expect(responses.broadcastTxCommitSuccess(txRes)).toEqual(true);
+      await tendermintSearchIndexUpdated();
+
+      // txSearch - you must enable the indexer when running
+      // tendermint, else you get empty results
+      const query = buildQuery({ tags: [{ key: "app.key", value: txKey2 }] });
+
+      const search = await client.txSearch({ query: query, page: 1, per_page: 30 });
+      // should find the tx
+      expect(search.totalCount).toEqual(1);
+      // should return same info as querying directly,
+      // except without the proof
+      expect(search.txs[0]).toEqual({
+        hash: txRes.hash,
+        height: txRes.height,
+        index: 0,
+        tx: tx,
+        result: txRes.deliverTx!,
+        proof: undefined,
+      });
+
+      // Ensure txSearchAll works as well. This should be moved in a dedicated "txSearchAll" test block.
+      const searchAll = await client.txSearchAll({ query: query });
+      expect(searchAll.totalCount).toEqual(1);
+      expect(searchAll.txs[0]).toEqual({
+        hash: txRes.hash,
+        height: txRes.height,
+        index: 0,
+        tx: tx,
+        result: txRes.deliverTx!,
+        proof: undefined,
+      });
+    });
+
+    it("returns transactions in ascending order by default", async () => {
       // NOTE: The Tendermint docs states the default ordering is "desc". Until
       // 0.35 it was actually "asc" but from 0.35 on it is "desc".
+      // Then it was changed back to "asc" in 0.37.
       // Docs: https://docs.tendermint.com/master/rpc/#/Info/tx_search
       // Code 0.34: https://github.com/tendermint/tendermint/blob/v0.34.10/rpc/core/tx.go#L89
       // Code 0.35: https://github.com/tendermint/tendermint/blob/v0.35.6/internal/rpc/core/tx.go#L93
+      // Code 0.37: https://github.com/cometbft/cometbft/blob/v0.37.0-rc3/rpc/core/tx.go#L87
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
 
-      const query = buildQuery({ tags: [{ key: "app.key", value: key }] });
+      const query = buildQuery({ tags: [{ key: "app.key", value: txKey }] });
 
       const result = await client.txSearch({ query: query });
 
       expect(result.totalCount).toEqual(3);
       result.txs.slice(1).reduce((lastHeight, { height }) => {
-        expect(height).toBeLessThanOrEqual(lastHeight);
+        expect(height).toBeGreaterThanOrEqual(lastHeight);
         return height;
       }, result.txs[0].height);
 
@@ -517,9 +565,9 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
 
     it("can set the order", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
 
-      const query = buildQuery({ tags: [{ key: "app.key", value: key }] });
+      const query = buildQuery({ tags: [{ key: "app.key", value: txKey }] });
 
       const result1 = await client.txSearch({ query: query, order_by: "desc" });
       const result2 = await client.txSearch({ query: query, order_by: "asc" });
@@ -532,9 +580,9 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
 
     it("can paginate over txSearch results", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
 
-      const query = buildQuery({ tags: [{ key: "app.key", value: key }] });
+      const query = buildQuery({ tags: [{ key: "app.key", value: txKey }] });
 
       // expect one page of results
       const s1 = await client.txSearch({ query: query, page: 1, per_page: 2 });
@@ -551,17 +599,16 @@ function defaultTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValues)
 
     it("can get all search results in one call", async () => {
       pendingWithoutTendermint();
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
 
-      const query = buildQuery({ tags: [{ key: "app.key", value: key }] });
+      const query = buildQuery({ tags: [{ key: "app.key", value: txKey }] });
 
       const sall = await client.txSearchAll({ query: query, per_page: 2 });
       expect(sall.totalCount).toEqual(3);
       expect(sall.txs.length).toEqual(3);
       // make sure there are in order from highest to lowest height
-      const [tx1, tx2, tx3] = sall.txs;
-      expect(tx2.height).toBeLessThan(tx1.height);
-      expect(tx3.height).toBeLessThan(tx2.height);
+      expect(sall.txs[1].height).toEqual(sall.txs[0].height + 1);
+      expect(sall.txs[2].height).toEqual(sall.txs[1].height + 1);
 
       client.disconnect();
     });
@@ -576,7 +623,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValue
 
     (async () => {
       const events: responses.NewBlockHeaderEvent[] = [];
-      const client = await Tendermint35Client.create(rpcFactory());
+      const client = await Tendermint37Client.create(rpcFactory());
       const stream = client.subscribeNewBlockHeader();
       expect(stream).toBeTruthy();
       const subscription = stream.subscribe({
@@ -635,7 +682,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValue
     const transactionData2 = buildKvTx(randomString(), randomString());
 
     const events: responses.NewBlockEvent[] = [];
-    const client = await Tendermint35Client.create(rpcFactory());
+    const client = await Tendermint37Client.create(rpcFactory());
     const stream = client.subscribeNewBlock();
     const subscription = stream.subscribe({
       next: (event) => {
@@ -694,7 +741,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValue
     pendingWithoutTendermint();
 
     const events: responses.TxEvent[] = [];
-    const client = await Tendermint35Client.create(rpcFactory());
+    const client = await Tendermint37Client.create(rpcFactory());
     const stream = client.subscribeTx();
     const subscription = stream.subscribe({
       next: (event) => {
@@ -738,7 +785,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValue
     const transactionData2 = buildKvTx(randomString(), randomString());
 
     const events: responses.TxEvent[] = [];
-    const client = await Tendermint35Client.create(rpcFactory());
+    const client = await Tendermint37Client.create(rpcFactory());
     const query = buildQuery({ tags: [{ key: "app.creator", value: expected.appCreator }] });
     const stream = client.subscribeTx(query);
     expect(stream).toBeTruthy();
@@ -776,7 +823,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValue
   it("can unsubscribe and re-subscribe to the same stream", async () => {
     pendingWithoutTendermint();
 
-    const client = await Tendermint35Client.create(rpcFactory());
+    const client = await Tendermint37Client.create(rpcFactory());
     const stream = client.subscribeNewBlockHeader();
 
     const event1 = await firstEvent(stream);
@@ -809,7 +856,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValue
   it("can subscribe twice", async () => {
     pendingWithoutTendermint();
 
-    const client = await Tendermint35Client.create(rpcFactory());
+    const client = await Tendermint37Client.create(rpcFactory());
     const stream1 = client.subscribeNewBlockHeader();
     const stream2 = client.subscribeNewBlockHeader();
 
@@ -821,15 +868,15 @@ function websocketTestSuite(rpcFactory: () => RpcClient, expected: ExpectedValue
   });
 }
 
-describe("Tendermint35Client", () => {
-  const { url, expected } = tendermintInstances[35];
+describe("Tendermint37Client", () => {
+  const { url, expected } = tendermintInstances[37];
 
   it("can connect to a given url", async () => {
     pendingWithoutTendermint();
 
     // default connection
     {
-      const client = await Tendermint35Client.connect(url);
+      const client = await Tendermint37Client.connect(url);
       const info = await client.abciInfo();
       expect(info).toBeTruthy();
       client.disconnect();
@@ -837,7 +884,7 @@ describe("Tendermint35Client", () => {
 
     // http connection
     {
-      const client = await Tendermint35Client.connect("http://" + url);
+      const client = await Tendermint37Client.connect("http://" + url);
       const info = await client.abciInfo();
       expect(info).toBeTruthy();
       client.disconnect();
@@ -845,7 +892,7 @@ describe("Tendermint35Client", () => {
 
     // ws connection
     {
-      const client = await Tendermint35Client.connect("ws://" + url);
+      const client = await Tendermint37Client.connect("ws://" + url);
       const info = await client.abciInfo();
       expect(info).toBeTruthy();
       client.disconnect();