+++ packages/tendermint-rpc/src/testutil.spec.ts
@@ -49,13 +49,13 @@ export const tendermintInstances = {
       appVersion: 1,
     },
   },
-  35: {
-    url: "localhost:11135",
-    version: "0.35.x",
+  37: {
+    url: "localhost:11137",
+    version: "0.37.x",
     blockTime: 500,
     expected: {
       chainId: /^dockerchain$/,
-      version: /^$/, // Unfortunately we don't get info here
+      version: /^0\.37\.0-alpha\.3$/,
       appCreator: "Cosmoshi Netowoko",
       p2pVersion: 8,
       blockVersion: 11,